+++ lib/MIP/Recipes/Analysis/Gatk_genotypegvcfs.pm
@@ -138,7 +138,7 @@ sub analysis_gatk_genotypegvcfs {
     use MIP::Pedigree qw{ create_fam_file };
     use MIP::Get::File qw{ get_io_files };
     use MIP::Get::Parameter qw{ get_recipe_attributes get_recipe_resources };
-    use MIP::Gnu::Coreutils qw{ gnu_cat gnu_echo gnu_rm };
+    use MIP::Program::Gnu::Coreutils qw{ gnu_cat gnu_echo gnu_rm };
     use MIP::Parse::File qw{ parse_io_outfiles };
     use MIP::Processmanagement::Processes qw{ submit_recipe };
     use MIP::Program::Gatk qw{ gatk_genomicsdbimport  gatk_genotypegvcfs };