+++ lib/MIP/Recipes/Analysis/Gatk_splitncigarreads.pm
@@ -158,7 +158,7 @@ sub analysis_gatk_splitncigarreads {
     use MIP::Cluster qw{ get_parallel_processes };
     use MIP::Get::File qw{ get_io_files };
     use MIP::Get::Parameter qw{ get_recipe_attributes get_recipe_resources };
-    use MIP::Gnu::Coreutils qw{ gnu_cp };
+    use MIP::Program::Gnu::Coreutils qw{ gnu_cp };
     use MIP::Parse::File qw{ parse_io_outfiles };
     use MIP::Processmanagement::Processes qw{ submit_recipe };
     use MIP::Program::Gatk qw{ gatk_splitncigarreads };