+++ t/gnu_cd.t
@@ -67,7 +67,7 @@ BEGIN {
     }
 
 ##Modules
-    my @modules = ('MIP::Gnu::Bash');
+    my @modules = ('MIP::Program::Gnu::Bash');
 
     for my $module (@modules) {
 
@@ -75,10 +75,10 @@ BEGIN {
     }
 }
 
-use MIP::Gnu::Bash qw(gnu_cd);
+use MIP::Program::Gnu::Bash qw(gnu_cd);
 use MIP::Test::Commands qw(test_function);
 
-diag("Test gnu_cd $MIP::Gnu::Bash::VERSION, Perl $^V, $EXECUTABLE_NAME");
+diag("Test gnu_cd $MIP::Program::Gnu::Bash::VERSION, Perl $^V, $EXECUTABLE_NAME");
 
 ## Base arguments
 my @function_base_commands = 'cd';