+++ t/gnu_ulimit.t
@@ -41,17 +41,17 @@ BEGIN {
 ### Check all internal dependency modules and imports
 ## Modules with import
     my %perl_module = (
-        q{MIP::Gnu::Bash}      => [qw{ gnu_ulimit }],
-        q{MIP::Test::Fixtures} => [qw{ test_standard_cli }],
+        q{MIP::Program::Gnu::Bash} => [qw{ gnu_ulimit }],
+        q{MIP::Test::Fixtures}     => [qw{ test_standard_cli }],
     );
 
     test_import( { perl_module_href => \%perl_module, } );
 }
 
-use MIP::Gnu::Bash qw{ gnu_ulimit };
+use MIP::Program::Gnu::Bash qw{ gnu_ulimit };
 
 diag(   q{Test gnu_ulimit from Bash.pm v}
-      . $MIP::Gnu::Bash::VERSION
+      . $MIP::Program::Gnu::Bash::VERSION
       . $COMMA
       . $SPACE . q{Perl}
       . $SPACE